add member func for opinfo

+++ paddle/ir/builtin_attribute_storage.cc
@@ -60,9 +60,9 @@ DictionaryAttributeStorage::ParamKey DictionaryAttributeStorage::GetAsKey()
 
 Attribute DictionaryAttributeStorage::GetValue(const StrAttribute &name) const {
   if (size_ > 0) {
-    size_t left = 0;
-    size_t right = size_ - 1;
-    size_t mid = 0;
+    int left = 0;
+    int right = size_ - 1;
+    int mid = 0;
     while (left <= right) {
       mid = (left + right) / 2;
       if (data_[mid].name() == name) {
